perf(brep): hoist the CartesianPoint cache across a worker's parts + per-part instrumentation
The native batch loop meshed each element on a fresh decoder, so every
faceted brep re-parsed its shared CartesianPoint lists from scratch. Switch
the rayon batch from flat_map_iter to map_init(FxHashMap::default, ...) so each
worker owns ONE point cache reused across its whole sub-range: process_entity_job
moves the warm cache into its local decoder (set_point_cache) and back out
(take_point_cache) per part. Job order is preserved (map_init + flatten_iter),
so mesh output is byte-identical (mesh_determinism no-diff).

Fold per-part point-cache hits/misses (always) and faceted-brep wall time
(observability + native only; Instant traps on wasm32) into request-local
atomics surfaced on ProcessingStats and PipelineDiagnostics. The wasm batch
already reuses one decoder, so it just reads decoder.point_cache_stats() at
batch end. PipelineDiagnostics crosses to JS as JsValue (typed any), so no
.d.ts change.

Extract process_entity_job + build_color_updates_for_jobs into a new
processor/jobs.rs sibling (module-size ratchet).
