+++ rust/processing/src/pipeline_diagnostics.rs
@@ -97,6 +97,16 @@ pub struct PipelineDiagnostics {
     pub silent_no_ops: u64,
     /// rect_fast fast-path engagement, summed across batches.
     pub rect_fast: RectFastSummary,
+    /// CartesianPoints served from the point cache while meshing faceted breps,
+    /// summed across batches. Non-zero proves cross-element point memoization
+    /// fired (the per-worker cache hoist). `hits / (hits + misses)` is the rate.
+    pub point_cache_hits: u64,
+    /// CartesianPoints parsed fresh (point-cache misses) across the faceted-brep
+    /// pass, summed across batches.
+    pub point_cache_misses: u64,
+    /// Wall time (ms) attributed to faceted-brep part meshing, summed across
+    /// batches. Populated only in `observability` native builds; 0 on wasm.
+    pub faceted_brep_ms: u64,
     /// Phase wall-times (native: full; wasm: geometry only).
     pub phase_ms: PipelinePhaseTimings,
 }
@@ -115,6 +125,9 @@ impl Default for PipelineDiagnostics {
             hosts_with_openings: 0,
             silent_no_ops: 0,
             rect_fast: RectFastSummary::default(),
+            point_cache_hits: 0,
+            point_cache_misses: 0,
+            faceted_brep_ms: 0,
             phase_ms: PipelinePhaseTimings::default(),
         }
     }
@@ -133,6 +146,9 @@ impl PipelineDiagnostics {
         mesh_count: u64,
         triangle_count: u64,
         backstop_count: u64,
+        point_cache_hits: u64,
+        point_cache_misses: u64,
+        faceted_brep_ms: u64,
         geometry_ms: u64,
         diag: &GeometryDiagnostics,
     ) {
@@ -141,6 +157,9 @@ impl PipelineDiagnostics {
         self.mesh_count += mesh_count;
         self.triangle_count += triangle_count;
         self.backstop_count += backstop_count;
+        self.point_cache_hits += point_cache_hits;
+        self.point_cache_misses += point_cache_misses;
+        self.faceted_brep_ms += faceted_brep_ms;
         self.total_csg_failures += diag.total_csg_failures;
         self.products_with_failures += diag.products_with_failures;
         self.hosts_with_openings += diag.hosts_with_openings;
@@ -179,6 +198,9 @@ impl PipelineDiagnostics {
             hosts_with_openings,
             silent_no_ops,
             rect_fast,
+            point_cache_hits: stats.point_cache_hits,
+            point_cache_misses: stats.point_cache_misses,
+            faceted_brep_ms: stats.faceted_brep_time_ms,
             phase_ms: PipelinePhaseTimings {
                 entity_scan_ms: stats.entity_scan_time_ms,
                 lookup_ms: stats.lookup_time_ms,
@@ -208,7 +230,7 @@ mod tests {
         // this JSON key set is what crosses to JS. Mirrors the
         // GeometryDiagnostics contract test in the geometry crate.
         let mut d = PipelineDiagnostics::default();
-        d.record_batch(10, 12, 3000, 2, 42, &GeometryDiagnostics::default());
+        d.record_batch(10, 12, 3000, 2, 40, 8, 5, 42, &GeometryDiagnostics::default());
         let v = serde_json::to_value(&d).expect("serializes");
         for key in [
             "schemaVersion",
@@ -222,6 +244,9 @@ mod tests {
             "hostsWithOpenings",
             "silentNoOps",
             "rectFast",
+            "pointCacheHits",
+            "pointCacheMisses",
+            "facetedBrepMs",
             "phaseMs",
         ] {
             assert!(v.get(key).is_some(), "missing top-level key {key}");
@@ -249,8 +274,8 @@ mod tests {
             hosts_with_openings: 3,
             ..Default::default()
         };
-        d.record_batch(10, 12, 3000, 1, 40, &diag);
-        d.record_batch(5, 6, 1500, 0, 10, &GeometryDiagnostics::default());
+        d.record_batch(10, 12, 3000, 1, 30, 4, 3, 40, &diag);
+        d.record_batch(5, 6, 1500, 0, 12, 1, 2, 10, &GeometryDiagnostics::default());
         assert!(!d.is_empty());
         assert_eq!(d.batches, 2);
         assert_eq!(d.element_count, 15);
@@ -259,6 +284,9 @@ mod tests {
         assert_eq!(d.backstop_count, 1);
         assert_eq!(d.total_csg_failures, 2);
         assert_eq!(d.hosts_with_openings, 3);
+        assert_eq!(d.point_cache_hits, 42);
+        assert_eq!(d.point_cache_misses, 5);
+        assert_eq!(d.faceted_brep_ms, 5);
         assert_eq!(d.phase_ms.geometry_ms, 50);
     }
 
@@ -276,6 +304,9 @@ mod tests {
             degenerate_triangles_dropped: 4,
             total_csg_failures: 1,
             products_with_failures: 1,
+            point_cache_hits: 128,
+            point_cache_misses: 32,
+            faceted_brep_time_ms: 9,
             ..Default::default()
         };
         let d = PipelineDiagnostics::from_processing_stats(&stats, 20);
@@ -283,6 +314,9 @@ mod tests {
         assert_eq!(d.element_count, 20);
         assert_eq!(d.mesh_count, 7);
         assert_eq!(d.backstop_count, 4);
+        assert_eq!(d.point_cache_hits, 128);
+        assert_eq!(d.point_cache_misses, 32);
+        assert_eq!(d.faceted_brep_ms, 9);
         assert_eq!(d.phase_ms.parse_ms, 11);
         assert_eq!(d.phase_ms.entity_scan_ms, 5);
         assert_eq!(d.phase_ms.lookup_ms, 2);
