+++ rust/processing/src/processor/jobs.rs
@@ -1,0 +1,284 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Per-entity geometry job execution.
+//!
+//! Split out of `processor/mod.rs` (module-size ratchet). `process_entity_job`
+//! is the body of the parallel batch loop's `map_init` closure — one job = one
+//! product, meshed on a fresh router with the worker's warm CartesianPoint
+//! cache moved in and back out. `build_color_updates_for_jobs` backfills
+//! deferred/orphan-type colours.
+
+use super::*;
+
+// Carries the full per-job processing context; factoring the args into a struct
+// would not change behavior and is out of scope for the lint gate.
+#[allow(clippy::too_many_arguments)]
+pub(super) fn process_entity_job(
+    job: &EntityJob,
+    content: &[u8],
+    entity_index_arc: &Arc<EntityIndex>,
+    unit_scale: f64,
+    rtc_offset: (f64, f64, f64),
+    // Pre-resolved scales seeded into this job's decoder so arc tessellation and
+    // unit conversion never trigger a per-element full-file IFCPROJECT scan.
+    seed_plane_angle_to_radians: f64,
+    tessellation_quality: TessellationQuality,
+    void_index: &FxHashMap<u32, Vec<u32>>,
+    skipped_entity_ids: &HashSet<u32>,
+    geometry_style_index: &FxHashMap<u32, GeometryStyleInfo>,
+    indexed_colour_full: &FxHashMap<u32, crate::style::FullIndexedColourMap>,
+    element_material_colors: &FxHashMap<u32, Vec<[f32; 4]>>,
+    // Surface textures + UV maps keyed by face-set id (#961). Empty for
+    // untextured models.
+    texture_index: &FxHashMap<u32, ifc_lite_geometry::ResolvedTextureMap>,
+    // Present only when the selected coordinate space is `site_local`; rotates
+    // mesh vertices into the site's axis frame.
+    site_local_rotation: Option<&Vec<f64>>,
+    // Shared sink for per-job router CSG diagnostics (parity with the wasm
+    // path's `drain_and_log_csg_diagnostics`).
+    csg_failure_collector: &std::sync::Mutex<FxHashMap<u32, Vec<ifc_lite_geometry::BoolFailure>>>,
+    // Shared sinks for opening classification + per-host opening diagnostics, drained
+    // from this job's router so the native pass aggregates the full GeometryDiagnostics.
+    classification_collector: &std::sync::Mutex<ifc_lite_geometry::ClassificationStats>,
+    host_diag_collector: &std::sync::Mutex<FxHashMap<u32, ifc_lite_geometry::HostOpeningDiagnostic>>,
+    rect_fast_collector: &std::sync::Mutex<ifc_lite_geometry::RectFastStats>,
+    // Shared tally of degenerate-backstop triangle drops (see
+    // `element::build_mesh_data`); relaxed atomic, added to only when non-zero.
+    backstop_collector: &std::sync::atomic::AtomicU64,
+    // Model-wide content-dedup cache shared by every per-job router so identical
+    // geometry is meshed once across the rayon pool (#1109 follow-up).
+    item_dedup_cache: &ifc_lite_geometry::ItemDedupCache,
+    // Per-WORKER CartesianPoint cache, owned by one rayon worker and reused across
+    // every element in its sub-range (see the `map_init` call site). Moved into
+    // this job's decoder and moved back out afterwards so a shared point list is
+    // parsed once per worker-batch, not once per part.
+    worker_point_cache: &mut FxHashMap<u32, (f64, f64, f64)>,
+    // Request-local point-cache hit/miss + faceted-brep-timing sinks (see the
+    // collectors declared before the chunk loop).
+    point_cache_hits_collector: &std::sync::atomic::AtomicU64,
+    point_cache_misses_collector: &std::sync::atomic::AtomicU64,
+    faceted_brep_ns_collector: &std::sync::atomic::AtomicU64,
+) -> Vec<MeshData> {
+    if skipped_entity_ids.contains(&job.id) {
+        return Vec::new();
+    }
+
+    let mut local_decoder = EntityDecoder::with_arc_index(content, entity_index_arc.clone());
+    // Adopt this worker's warm point cache so faceted-brep polyloop points shared
+    // across elements are served from memory instead of re-parsed per element.
+    local_decoder.set_point_cache(std::mem::take(worker_point_cache));
+    // Seed the unit-scale caches so curve/arc processing skips the O(file)
+    // IFCPROJECT scan that each fresh per-element decoder would otherwise repeat.
+    local_decoder.seed_unit_scales(unit_scale, seed_plane_angle_to_radians);
+
+    let entity = match local_decoder.decode_at(job.start, job.end) {
+        Ok(entity) => entity,
+        Err(_) => return Vec::new(),
+    };
+
+    let mut local_router = GeometryRouter::with_scale_and_quality(unit_scale, tessellation_quality);
+    local_router.set_rtc_offset(rtc_offset);
+    // content-dedup default OFF: its structural hash costs more than the meshing
+    // it skips on real models (see GeometryRouter::content_dedup_enabled).
+    if GeometryRouter::content_dedup_enabled() {
+        local_router.enable_content_dedup_shared(item_dedup_cache.clone());
+    }
+    let local_router = local_router;
+
+    let metadata = crate::element::ElementMeshMetadata {
+        global_id: job.global_id.clone(),
+        name: job.name.clone(),
+        presentation_layer: job.presentation_layer.clone(),
+        space_zone_properties: job.space_zone_properties.clone(),
+    };
+    // #957: the scan loop plans type geometry with `SuppressInstanced` (see
+    // `plan_type_geometry`), so a synthetic job's map always renders as an
+    // orphan — geometry_class 1.
+    let kind = match job.representation_map_id {
+        Some(rep_map_id) => crate::element::ElementJobKind::TypeProduct {
+            rep_maps: vec![(rep_map_id, 1)],
+        },
+        None => crate::element::ElementJobKind::Product,
+    };
+    let ctx = crate::element::MeshProductionContext {
+        void_index,
+        geometry_style_index,
+        indexed_colour_full,
+        element_material_colors,
+        texture_index,
+        site_local_rotation,
+    };
+
+    // Per-part faceted-brep phase timing (observability only; Instant traps on
+    // wasm32). Attributed below to `faceted_brep_ns_collector` for parts that
+    // actually touched the point cache (i.e. faceted breps), mirroring the
+    // diag.rs feature-off byte-identity policy.
+    #[cfg(all(feature = "observability", not(target_arch = "wasm32")))]
+    let brep_timer = std::time::Instant::now();
+
+    let produced = crate::element::produce_element_meshes(
+        &crate::element::ElementMeshJob {
+            id: job.id,
+            ifc_type: job.ifc_type,
+            entity: &entity,
+            kind,
+            element_color: Some(job.element_color),
+            metadata: Some(&metadata),
+        },
+        &ctx,
+        // Geometry hashing is a viewer diff feature — off on the native path.
+        &crate::element::MeshProductionOptions::default(),
+        &mut local_decoder,
+        &local_router,
+    );
+
+    // Fold this element's point-cache activity into the pass tallies, then move
+    // the warm cache back to the worker for the next element. `hits + misses > 0`
+    // means this element was a faceted brep that walked polyloops.
+    let (part_hits, part_misses) = local_decoder.point_cache_stats();
+    if part_hits + part_misses > 0 {
+        use std::sync::atomic::Ordering::Relaxed;
+        point_cache_hits_collector.fetch_add(part_hits, Relaxed);
+        point_cache_misses_collector.fetch_add(part_misses, Relaxed);
+        // Faceted-brep wall time is measured only under `observability` (Instant
+        // traps on wasm32); the timer/log block compiles out entirely with the
+        // feature off, so default builds stay byte-identical. Adding a 0 ns delta
+        // in that case keeps `faceted_brep_ns_collector` always-used.
+        #[cfg(all(feature = "observability", not(target_arch = "wasm32")))]
+        {
+            let elapsed_ns = brep_timer.elapsed().as_nanos() as u64;
+            faceted_brep_ns_collector.fetch_add(elapsed_ns, Relaxed);
+            // Under `observability` this is what the geometry crate's `diag_debug!`
+            // expands to (`tracing::debug!`).
+            tracing::debug!(
+                element_id = job.id,
+                point_cache_hits = part_hits,
+                point_cache_misses = part_misses,
+                faceted_brep_us = elapsed_ns / 1000,
+                "faceted-brep part meshed"
+            );
+        }
+        #[cfg(not(all(feature = "observability", not(target_arch = "wasm32"))))]
+        faceted_brep_ns_collector.fetch_add(0, Relaxed);
+    }
+    *worker_point_cache = local_decoder.take_point_cache();
+
+    // Fold this element's degenerate-backstop drops into the pass tally.
+    if produced.degenerate_triangles_dropped > 0 {
+        backstop_collector.fetch_add(
+            produced.degenerate_triangles_dropped,
+            std::sync::atomic::Ordering::Relaxed,
+        );
+    }
+
+    // Surface this element's CSG diagnostics in the shared collector. The
+    // wasm path logs them in the browser console; without this the server
+    // would silently discard every failed opening cut.
+    if !produced.csg_failures.is_empty() {
+        if let Ok(mut collector) = csg_failure_collector.lock() {
+            for (product_id, fails) in produced.csg_failures {
+                collector.entry(product_id).or_default().extend(fails);
+            }
+        }
+    }
+
+    // Drain this job's router opening diagnostics (classification counts +
+    // per-host detail) and merge them. Each job uses a FRESH router that
+    // processed exactly this one element, so the drained values are this
+    // element's only — summing across jobs mirrors the wasm batch router's
+    // accumulate-then-drain, giving the same GeometryDiagnostics counts.
+    let cls = local_router.take_classification_stats();
+    if cls.rectangular != 0
+        || cls.diagonal != 0
+        || cls.non_rectangular != 0
+        || cls.floor_opening_guard_saved != 0
+    {
+        if let Ok(mut acc) = classification_collector.lock() {
+            acc.rectangular += cls.rectangular;
+            acc.diagonal += cls.diagonal;
+            acc.non_rectangular += cls.non_rectangular;
+            acc.floor_opening_guard_saved += cls.floor_opening_guard_saved;
+        }
+    }
+    let host_diags = local_router.take_host_opening_diagnostics();
+    if !host_diags.is_empty() {
+        if let Ok(mut acc) = host_diag_collector.lock() {
+            // Product ids are disjoint across jobs (one product = one job), so this
+            // is an insert; `extend` is robust if that ever changes.
+            acc.extend(host_diags);
+        }
+    }
+    // Drain this job's router rect_fast counters into the request-local collector
+    // (process-global counters are gone — see GeometryRouter::record_rect_fast).
+    let rf = local_router.take_rect_fast_stats();
+    if rf.fired != 0
+        || rf.openings_cut != 0
+        || rf.defer_host_not_box != 0
+        || rf.defer_not_through != 0
+        || rf.defer_off_face != 0
+        || rf.defer_near_edge != 0
+        || rf.defer_no_openings != 0
+    {
+        if let Ok(mut acc) = rect_fast_collector.lock() {
+            acc.fired += rf.fired;
+            acc.openings_cut += rf.openings_cut;
+            acc.defer_host_not_box += rf.defer_host_not_box;
+            acc.defer_not_through += rf.defer_not_through;
+            acc.defer_off_face += rf.defer_off_face;
+            acc.defer_near_edge += rf.defer_near_edge;
+            acc.defer_no_openings += rf.defer_no_openings;
+        }
+    }
+
+    produced.meshes
+}
+
+pub(super) fn build_color_updates_for_jobs(
+    jobs: &[EntityJob],
+    geometry_styles: &FxHashMap<u32, GeometryStyleInfo>,
+    content: &[u8],
+    entity_index: &Arc<EntityIndex>,
+) -> Vec<(u32, [f32; 4])> {
+    let mut decoder = EntityDecoder::with_arc_index(content, entity_index.clone());
+    let mut updates: Vec<(u32, [f32; 4])> = Vec::new();
+
+    for job in jobs {
+        // #957: synthetic type-only-geometry jobs resolve their colour from the
+        // RepresentationMap (a type has no IfcProductDefinitionShape), so the
+        // product-definition path below never corrects them. Backfill them here
+        // or a deferred IfcStyledItem (fast_first_batch) leaves the orphan type
+        // geometry stuck at its fallback colour.
+        if let Some(rep_map_id) = job.representation_map_id {
+            if let Some(color) = crate::element::resolve_color_for_representation_map(
+                rep_map_id,
+                geometry_styles,
+                &mut decoder,
+            ) {
+                if color != job.element_color {
+                    updates.push((job.id, color));
+                }
+            }
+            continue;
+        }
+        let Ok(entity) = decoder.decode_at(job.start, job.end) else {
+            continue;
+        };
+        let Some(product_definition_shape_id) = entity.get_ref(6) else {
+            continue;
+        };
+        let Some(color) = resolve_element_color_for_product_definition_shape(
+            product_definition_shape_id,
+            geometry_styles,
+            &mut decoder,
+        ) else {
+            continue;
+        };
+        if color != job.element_color {
+            updates.push((job.id, color));
+        }
+    }
+
+    updates
+}
